Move DataTool to src/components
