+++ src/components/App/App.js
@@ -4,11 +4,11 @@ import React from 'react';
 import { Navbar } from 'react-bootstrap';
 import { BrowserRouter as Router } from 'react-router-dom';
 
-import DataTool from '../navigation/DataTool';
+import DataTool from '../DataTool';
 import NavRoutes from '../navigation/NavRoutes';
-import Help from '../navigation/Help';
+import Help from '../help-etc/Help';
 import Science from '../help-etc/Science';
-import About from '../navigation/About';
+import About from '../help-etc/About';
 
 import logo from '../../assets/logo.png';
 import marmot from '../../assets/marmot.png';
